Split AppWrapper controller into separate dispatcher and runner
Progress towards multi-cluster MCAD by splitting the AppWrapper
reconciler into two separate reconcilers that handle disjoint
state transitions in the lifecycle of the AppWrapper.
