+++ test/e2e-kuttl/steps/06-assert.yaml
@@ -7,8 +7,14 @@ status:
   transitions:
   - state: Pending
   - state: Running
+    step: dispatching
+  - state: Running
+    step: accepted
+  - state: Running
     step: creating
   - state: Running
     step: created
+  - state: Completed
+    step: returned
   - state: Completed
   state: Completed
