+++ test/e2e-kuttl/steps/08-assert.yaml
@@ -7,6 +7,10 @@ status:
   transitions:
   - state: Pending
   - state: Running
+    step: dispatching
+  - state: Running
+    step: accepted
+  - state: Running
     step: creating
   - state: Running
     step: created
